服务器列表增加 CPU使用率、内存使用率、上行带宽、下行带宽

+++ monitoring-ui/src/main/java/com/gitee/pifeng/monitoring/ui/business/web/service/impl/MonitorServerNetcardServiceImpl.java
@@ -1,9 +1,9 @@
 package com.gitee.pifeng.monitoring.ui.business.web.service.impl;
 
-import cn.hutool.core.io.unit.DataSizeUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.gitee.pifeng.monitoring.common.constant.ZeroOrOneConstants;
+import com.gitee.pifeng.monitoring.common.util.DataSizeUtil;
 import com.gitee.pifeng.monitoring.ui.business.web.dao.IMonitorServerNetcardDao;
 import com.gitee.pifeng.monitoring.ui.business.web.entity.MonitorServerNetcard;
 import com.gitee.pifeng.monitoring.ui.business.web.service.IMonitorServerNetcardService;
@@ -69,8 +69,8 @@ public class MonitorServerNetcardServiceImpl extends ServiceImpl<IMonitorServerN
             serverDetailPageServerNetcardVo.setTxPackets(monitorServerNetcard.getTxPackets());
             serverDetailPageServerNetcardVo.setTxErrors(monitorServerNetcard.getTxErrors());
             serverDetailPageServerNetcardVo.setTxDropped(monitorServerNetcard.getTxDropped());
-            String downloadSpeed = DataSizeUtil.format(monitorServerNetcard.getDownloadBps().longValue());
-            String uploadSpeed = DataSizeUtil.format(monitorServerNetcard.getUploadBps().longValue());
+            String downloadSpeed = DataSizeUtil.format(monitorServerNetcard.getDownloadBps());
+            String uploadSpeed = DataSizeUtil.format(monitorServerNetcard.getUploadBps());
             serverDetailPageServerNetcardVo.setDownloadSpeed(ZeroOrOneConstants.ZERO.equals(downloadSpeed) ? ZeroOrOneConstants.ZERO : downloadSpeed + "/s");
             serverDetailPageServerNetcardVo.setUploadSpeed(ZeroOrOneConstants.ZERO.equals(uploadSpeed) ? ZeroOrOneConstants.ZERO : uploadSpeed + "/s");
             serverDetailPageServerNetcardVos.add(serverDetailPageServerNetcardVo);
